Small tidy up
